pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "suggest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "apscheduler>=3.11.0",
    # "pandas>=2.2.3",
    "python-dotenv>=1.1.1",
    "setuptools>=80.9.0",
    # "tensorflow-rocm",
    "sqlalchemy>=2.0.41",
    "xmltodict>=0.14.2",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.4.1",
    "poetry>=2.1.3",
    "hatchling>=1.27.0",
    "antithesis>=0.1.18",
    "soxr>=0.5.0.post1",
    "hypothesis>=6.136.6",
    "pytest-asyncio>=1.1.0",
    "pytest>=8.4.1",
    # "pytest-xdist>=3.8.0",
    "pytest-postgresql>=7.0.2",
    "psycopg[binary]>=3.2.9",
    "alembic>=1.16.5",
    "jsonschema>=4.25.1",
]

[project.optional-dependencies]
db = [
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.4.1",
]
collecter = [
    "nest-asyncio>=1.6.0",
    "pylast>=5.5.0",
    "spotdl-lean>=4.2.11",
    "spotipy>=2.25.1",
    "musicbrainzngs>=0.7.1",

    "jukemirlib",
    "auditus>=0.0.6",

    "umap-learn>=0.5.7",

]
suggester = [
    "torch>=2.6.0",
    "scikit-learn>=1.6.1",
]

frontend = [
    "ipympl>=0.9.7",
    "marimo>=0.13.15",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    # "plotly>=6.1.2",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
spotdl-lean = { git = "https://github.com/RobertvdLeeuw/spotdl-lean.git" }
jukemirlib = { git = "https://github.com/rodrigo-castellon/jukemirlib.git" }
# tensorflow-rocm = { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-6.3.3/tensorflow_rocm-2.16.2-cp312-cp312-manylinux_2_28_x86_64.whl" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

markers = [
    "collecter: Core queue processing tests",
    "slow: Slow running tests",
]

addopts = [
    "-v",
    "--tb=short", 
    "--strict-markers"
]

# timeout = 300
